Create command mvnRunClass.
Command mvnRunClass can run a class from a given package in repository.

For instance the command below, download the scala package if not
available in the cache and run the main class
scala.tools.nsc.MainGenericRunner which invokes its REPL.

 $ scala jarget.jar mvn-run-class org.scala-lang.virtualized/scala-compiler/2.11.2 \
   scala.tools.nsc.MainGenericRunner -Dscala.usejavacp=true

The package and its dependencies are downloaded only once. In the next
time this command is executed, they will be invoked from the jarget
cache directory.